F: date_time arithmetic overflow warning

+++ xo/time/date_time.cpp
@@ -6,11 +6,11 @@ namespace xo
 	{
 		return
 			uint64( year ) * 12 * 31 * 24 * 60 * 60 * 1000 +
-			month * 31 * 24 * 60 * 60 * 1000 +
-			( day - 1 ) * 24 * 60 * 60 * 1000 +
-			hour * 60 * 60 * 1000 +
-			minute * 60 * 1000 +
-			second * 1000 +
-			millisecond;
+			uint64(month )* 31 * 24 * 60 * 60 * 1000 +
+			( uint64( day ) - 1 ) * 24 * 60 * 60 * 1000 +
+			uint64(hour )* 60 * 60 * 1000 +
+			uint64(minute )* 60 * 1000 +
+			uint64(second )* 1000 +
+			uint64(millisecond);
 	}
 }
